backend/requirements.txt
fastapi
python-dotenv
uvicorn
google-generativeai
google-genai  
pydantic
python-multipart
pymupdf
pillow 
pytesseract